--- src/components/NavigationBar.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-function NavigationBar() {
-    return(
-        <div className="navigation-bar">
-            <button>끼니 PICK</button>
-            <button>끼니 랭킹</button>
-            <button>끼니 그린 랭킹</button>
-            <button>커뮤니티</button>
-        </div>
-    );
-}
-
-export default NavigationBar;
