+++ src/containers/MainPageContainer.js
@@ -1,20 +1,18 @@
 import React from 'react';
-import Header from "../components/Header";
-import NavigationBar from "../components/NavigationBar";
 import RankingSection from "../components/RankingSection";
+import SearchBar from "../components/SearchBar";
 import Footer from "../components/Footer";
 import BannerAd from "../components/BannerAd";
 
 function MainPageContainer({ searchTerm, setSearchTerm, autocompleteItems, setAutocompleteItems }) {
     return (
         <>
-            <Header
+            <SearchBar
                 searchTerm={searchTerm}
                 setSearchTerm={setSearchTerm}
                 autocompleteItems={autocompleteItems}
                 setAutocompleteItems={setAutocompleteItems}
             />
-            <NavigationBar/>
             <BannerAd/>
             <RankingSection/>
             <Footer/>
